+++ Content.Shared/_Goobstation/MisandryBox/Smites/InputSwapSystem.cs
@@ -1,0 +1,17 @@
+﻿using Content.Shared.Movement.Components;
+using Content.Shared.Movement.Systems;
+
+namespace Content.Shared._Goobstation.MisandryBox.Smites;
+
+public sealed class InputSwapSystem : ToggleableSmiteSystem<InputSwapComponent>
+{
+    [Dependency] private readonly MovementSpeedModifierSystem _move = default!;
+
+    public override void Set(EntityUid ent)
+    {
+        var mod = Comp<MovementSpeedModifierComponent>(ent);
+
+        _move.ChangeBaseSpeed(ent, -mod.BaseWalkSpeed, -mod.BaseSprintSpeed, mod.Acceleration);
+        _move.RefreshMovementSpeedModifiers(ent);
+    }
+}
